// frontend/src/lib/generated/api/models/DTOCurso.ts
export interface DTOCurso {
    /**
     * 
     * @type {number}
     * @memberof DTOCurso
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOCurso
     */
    titulo?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOCurso
     */
    descripcion?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOCurso
     */
    precio?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOCurso
     */
    presencialidad?: DTOCursoPresencialidadEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOCurso
     */
    imagenPortada?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOCurso
     */
    nivel?: DTOCursoNivelEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOCurso
     */
    alumnosId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOCurso
     */
    profesoresId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOCurso
     */
    ejerciciosId?: Array<string>;
    /**
     * 
     * @type {Array<DTOMaterial>}
     * @memberof DTOCurso
     */
    material?: Array<DTOMaterial>;
    /**
     * 
     * @type {Date}
     * @memberof DTOCurso
     */
    fechaInicio?: Date;
    /**
     * 
     * @type {Date}
     * @memberof DTOCurso
     */
    fechaFin?: Date;
    /**
     * 
     * @type {number}
     * @memberof DTOCurso
     */
    numeroProfesores?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOCurso
     */
    duracionEnDias?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOCurso
     */
    duracionEnSemanas?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOCurso
     */
    porcentajeProgreso?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOCurso
     */
    numeroAlumnos?: number;
}


/**
 * @export
 */
export const DTOCursoPresencialidadEnum = {
    Online: 'ONLINE',
    Presencial: 'PRESENCIAL',
    Hibrido: 'HIBRIDO'
} as const;
export type DTOCursoPresencialidadEnum = typeof DTOCursoPresencialidadEnum[keyof typeof DTOCursoPresencialidadEnum];

/**
 * @export
 */
export const DTOCursoNivelEnum = {
    Principiante: 'PRINCIPIANTE',
    Intermedio: 'INTERMEDIO',
    Avanzado: 'AVANZADO',
    Basico: 'BASICO'
} as const;
export type DTOCursoNivelEnum = typeof DTOCursoNivelEnum[keyof typeof DTOCursoNivelEnum];