// frontend/src/lib/generated/api/models/DTOTaller.ts
export interface DTOTaller {
    /**
     * 
     * @type {number}
     * @memberof DTOTaller
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOTaller
     */
    titulo?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOTaller
     */
    descripcion?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOTaller
     */
    precio?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOTaller
     */
    presencialidad?: DTOTallerPresencialidadEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOTaller
     */
    imagenPortada?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOTaller
     */
    nivel?: DTOTallerNivelEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOTaller
     */
    alumnosId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOTaller
     */
    profesoresId?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof DTOTaller
     */
    ejerciciosId?: Array<string>;
    /**
     * 
     * @type {Array<DTOMaterial>}
     * @memberof DTOTaller
     */
    material?: Array<DTOMaterial>;
    /**
     * 
     * @type {number}
     * @memberof DTOTaller
     */
    duracionHoras?: number;
    /**
     * 
     * @type {Date}
     * @memberof DTOTaller
     */
    fechaRealizacion?: Date;
    /**
     * 
     * @type {string}
     * @memberof DTOTaller
     */
    horaComienzo?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOTaller
     */
    numeroProfesores?: number;
    /**
     * 
     * @type {string}
     * @memberof DTOTaller
     */
    horaFinalizacion?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOTaller
     */
    numeroAlumnos?: number;
}


/**
 * @export
 */
export const DTOTallerPresencialidadEnum = {
    Online: 'ONLINE',
    Presencial: 'PRESENCIAL',
    Hibrido: 'HIBRIDO'
} as const;
export type DTOTallerPresencialidadEnum = typeof DTOTallerPresencialidadEnum[keyof typeof DTOTallerPresencialidadEnum];

/**
 * @export
 */
export const DTOTallerNivelEnum = {
    Principiante: 'PRINCIPIANTE',
    Intermedio: 'INTERMEDIO',
    Avanzado: 'AVANZADO',
    Basico: 'BASICO'
} as const;
export type DTOTallerNivelEnum = typeof DTOTallerNivelEnum[keyof typeof DTOTallerNivelEnum];