// frontend/src/lib/generated/api/models/DTOPeticionCrearCurso.ts
export interface DTOPeticionCrearCurso {
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearCurso
     */
    titulo: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearCurso
     */
    descripcion?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOPeticionCrearCurso
     */
    precio: number;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearCurso
     */
    presencialidad: DTOPeticionCrearCursoPresencialidadEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearCurso
     */
    imagenPortada?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearCurso
     */
    nivel: DTOPeticionCrearCursoNivelEnum;
    /**
     * 
     * @type {Date}
     * @memberof DTOPeticionCrearCurso
     */
    fechaInicio: Date;
    /**
     * 
     * @type {Date}
     * @memberof DTOPeticionCrearCurso
     */
    fechaFin: Date;
    /**
     * 
     * @type {Array<number>}
     * @memberof DTOPeticionCrearCurso
     */
    profesoresId?: Array<number>;
    /**
     * 
     * @type {Array<Material>}
     * @memberof DTOPeticionCrearCurso
     */
    material?: Array<Material>;
}


/**
 * @export
 */
export const DTOPeticionCrearCursoPresencialidadEnum = {
    Online: 'ONLINE',
    Presencial: 'PRESENCIAL',
    Hibrido: 'HIBRIDO'
} as const;
export type DTOPeticionCrearCursoPresencialidadEnum = typeof DTOPeticionCrearCursoPresencialidadEnum[keyof typeof DTOPeticionCrearCursoPresencialidadEnum];

/**
 * @export
 */
export const DTOPeticionCrearCursoNivelEnum = {
    Principiante: 'PRINCIPIANTE',
    Intermedio: 'INTERMEDIO',
    Avanzado: 'AVANZADO',
    Basico: 'BASICO'
} as const;
export type DTOPeticionCrearCursoNivelEnum = typeof DTOPeticionCrearCursoNivelEnum[keyof typeof DTOPeticionCrearCursoNivelEnum];