// frontend/src/lib/generated/api/models/DTOPeticionCrearTaller.ts
export interface DTOPeticionCrearTaller {
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearTaller
     */
    titulo: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearTaller
     */
    descripcion?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOPeticionCrearTaller
     */
    precio: number;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearTaller
     */
    presencialidad: DTOPeticionCrearTallerPresencialidadEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearTaller
     */
    imagenPortada?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearTaller
     */
    nivel: DTOPeticionCrearTallerNivelEnum;
    /**
     * 
     * @type {number}
     * @memberof DTOPeticionCrearTaller
     */
    duracionHoras: number;
    /**
     * 
     * @type {Date}
     * @memberof DTOPeticionCrearTaller
     */
    fechaRealizacion: Date;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionCrearTaller
     */
    horaComienzo: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof DTOPeticionCrearTaller
     */
    profesoresId?: Array<number>;
    /**
     * 
     * @type {Array<Material>}
     * @memberof DTOPeticionCrearTaller
     */
    material?: Array<Material>;
}


/**
 * @export
 */
export const DTOPeticionCrearTallerPresencialidadEnum = {
    Online: 'ONLINE',
    Presencial: 'PRESENCIAL',
    Hibrido: 'HIBRIDO'
} as const;
export type DTOPeticionCrearTallerPresencialidadEnum = typeof DTOPeticionCrearTallerPresencialidadEnum[keyof typeof DTOPeticionCrearTallerPresencialidadEnum];

/**
 * @export
 */
export const DTOPeticionCrearTallerNivelEnum = {
    Principiante: 'PRINCIPIANTE',
    Intermedio: 'INTERMEDIO',
    Avanzado: 'AVANZADO',
    Basico: 'BASICO'
} as const;
export type DTOPeticionCrearTallerNivelEnum = typeof DTOPeticionCrearTallerNivelEnum[keyof typeof DTOPeticionCrearTallerNivelEnum];